Rewrote cassandra session management
